refactor(small changes): Small fixes
